+++ src/CircularString.php
@@ -153,6 +153,9 @@ class CircularString extends Curve
         return $boundingBox;
     }
 
+    /**
+     * @return list<list<float>>
+     */
     #[Override]
     public function toArray() : array
     {