+++ src/CompoundCurve.php
@@ -161,6 +161,9 @@ class CompoundCurve extends Curve
         return $boundingBox;
     }
 
+    /**
+     * @return list<list<list<float>>>
+     */
     #[Override]
     public function toArray() : array
     {