+++ src/LineString.php
@@ -195,6 +195,9 @@ class LineString extends Curve
         return $boundingBox;
     }
 
+    /**
+     * @return list<list<float>>
+     */
     #[Override]
     public function toArray() : array
     {