+++ src/MultiLineString.php
@@ -15,6 +15,18 @@ use Override;
  */
 class MultiLineString extends MultiCurve
 {
+    /**
+     * @return list<list<list<float>>>
+     */
+    #[Override]
+    public function toArray() : array
+    {
+        return array_map(
+            fn(LineString $lineString) => $lineString->toArray(),
+            $this->geometries,
+        );
+    }
+
     #[NoProxy, Override]
     public function geometryType() : string
     {