+++ src/MultiPolygon.php
@@ -29,6 +29,18 @@ use Override;
  */
 class MultiPolygon extends MultiSurface
 {
+    /**
+     * @return list<list<list<list<float>>>>
+     */
+    #[Override]
+    public function toArray() : array
+    {
+        return array_map(
+            fn(Polygon $polygon) => $polygon->toArray(),
+            $this->geometries,
+        );
+    }
+
     #[NoProxy, Override]
     public function geometryType() : string
     {