+++ src/Point.php
@@ -231,9 +231,7 @@ class Point extends Geometry
     }
 
     /**
-     * @psalm-return list<float>
-     *
-     * @return float[]
+     * @return list<float>
      */
     #[Override]
     public function toArray() : array