+++ src/Polygon.php
@@ -178,6 +178,9 @@ class Polygon extends Surface
         return $boundingBox;
     }
 
+    /**
+     * @return list<list<list<float>>>
+     */
     #[Override]
     public function toArray() : array
     {