+++ src/PolyhedralSurface.php
@@ -137,6 +137,9 @@ class PolyhedralSurface extends Surface
         return $boundingBox;
     }
 
+    /**
+     * @return list<list<list<list<float>>>>
+     */
     #[Override]
     public function toArray() : array
     {